authentication added

+++ base/views.py
@@ -2,6 +2,12 @@ from django.shortcuts import render,redirect
 from .models import Room,Topic
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+
 
 # Create from django.http import HttpResponse
 
@@ -21,6 +27,7 @@ def room(request,pk):
     room=Room.objects.get(id=pk)
     context={'room':room}
     return render(request,'base/room.html',context)
+@login_required(login_url='login')
 def CreateRoom(request):
     form=RoomForm()
     if(request.method=='POST'):
@@ -30,10 +37,12 @@ def CreateRoom(request):
             return redirect('home')
     context={'form':form}
     return render(request,'base/room_form.html',context)
-
+@login_required(login_url='login')
 def UpdateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if(room.host!=request.user):
+        return HttpResponse("you're not allowed here my friend ")
     if(request.method=='POST'):
         form=RoomForm(request.POST,instance=room)
         if(form.is_valid()):
@@ -41,9 +50,44 @@ def UpdateRoom(request,pk):
             return redirect('home')
     context={'form':form}
     return render(request,'base/room_form.html',context)
+
+
+@login_required(login_url='login')
 def DeleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if(room.host!=request.user):
+        return HttpResponse("you're not allowed here my friend ")
     if request.method=='POST':
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+
+
+def LoginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method=="POST":
+        username=request.POST.get('username')
+        password=request.POST.get('password')
+        try:
+            user=User.objects.get(username=username)
+        except:
+            messages.error(request,"user does not exists")
+        user=authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)  
+            return redirect('home')
+        else:
+            messages.error(request,"password is false ..try again")
+    
+    context={'page':page}
+    return render(request,"base/login_register.html",context)
+def Logout(request):
+    logout(request)
+    return redirect('home')
+def registerPage(request):
+    page='register'
+    context={'page':page}
+    return redirect(request,'base/login_register.html',context)
